move tests to different class
